BatchNormalizer: Hypertrain/Markings/Characteristics added
v5.5.2

// .[MarkingKey]=[ColorVal] → Mark:
// Value is a series of marking assignments like "Markings: Diamond=Red / Heart=Blue / Circle=No"
// Any Markings that are not specified or include "No" are ignored

-------

// Creates a ".Characteristic=" batch command that can be written as "Characteristic:"
// Value is a characteristic wordage like "Characteristic: Likes to eat"
// Each characteristic has a predetermined IV set that guarantees that it's present on the Pokémon

-------

// .HT_[STAT]= → HT:
// "HT: HP / Atk / Def" to enable Hyper Training for those stats only

-------

If a user implements the "Characteristic:" command, but also uses the "IVs:" or "Set IVs" command, it will be overwritten by the predetermined Characteristic set and ignore the IV lines.

Markings will take priority over an IV that is 31, since the bot automatically fills in the Marking objects that align with whichever IV is perfect. It will be ignored for the Markings command.